feat(mag): keep strapdown compensation script beside the flight fit
Replacing mag_compensation.py in place dropped the restrained-ramp tool some people still want.
